# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cachet API",
    "description": "Issuance console and verifiable registry for Zcash Shielded Assets (ZSA), running against the public ZSA testnet.\n\nEverything the registry serves is **derived from the chain** and independently checkable: metadata hashes participate in the asset id itself (ZIP 227), so a response can be re-verified client-side. The public instance at `api.cachetzec.com` is **read-only** — every mutating `issuance` endpoint answers 403 there; browser mints go through `POST /api/v1/relay`, which only forwards signed bytes.\n\nRate limit: 30 requests/s per client, bursts of 60 - sized for a real session, a browser wallet paging through the chain included. Need better limits, or none at all? Run your own instance: the code is MIT, the limit is one environment variable (`CACHET_RATE_LIMIT_PER_SEC`, `0` disables it), and the public ZSA node accepts transactions directly. This deployment is a convenience, never a chokepoint.\n\nNo cookies, no telemetry, and no authentication anywhere except the optional operator surface.",
    "license": {
      "name": "MIT"
    },
    "version": "0.4.0"
  },
  "paths": {
    "/api/v1/admin/moderation": {
      "get": {
        "tags": ["ops"],
        "summary": "Every moderation entry, for the operator's audit view.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModerationEntryResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Admin surface disabled or bad token"
          }
        }
      },
      "post": {
        "tags": ["ops"],
        "summary": "Hide a bundle, description or issuer from this registry's distribution.",
        "operationId": "hide",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Hidden"
          },
          "404": {
            "description": "Admin surface disabled or bad token"
          }
        }
      },
      "delete": {
        "tags": ["ops"],
        "summary": "Lift a moderation entry.",
        "operationId": "unhide",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Unhidden (or was not hidden)"
          },
          "404": {
            "description": "Admin surface disabled or bad token"
          }
        }
      }
    },
    "/api/v1/admin/pause": {
      "get": {
        "tags": ["ops"],
        "summary": "The current pause decision.",
        "operationId": "pause_get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Admin surface disabled or bad token"
          }
        }
      },
      "put": {
        "tags": ["ops"],
        "summary": "Pause or resume minting through this instance. Effective on the next\nrequest; persisted so a restart keeps it. The chain is not involved:\na paused instance simply stops relaying and storing.",
        "operationId": "pause_set",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Admin surface disabled or bad token"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModerationEntryResponse": {
        "type": "object",
        "required": ["kind", "key", "hidden_at"],
        "properties": {
          "bytes_present": {
            "type": ["boolean", "null"],
            "description": "Bundles only: whether this registry still holds the bytes\n(`false` after a purge). `null` for other kinds."
          },
          "hidden_at": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "reason": {
            "type": ["string", "null"]
          }
        }
      },
      "ModerationRequest": {
        "type": "object",
        "required": ["kind", "key"],
        "properties": {
          "key": {
            "type": "string",
            "description": "Hex-encoded key: bundle sha256, asset id, or issuance validating key."
          },
          "kind": {
            "type": "string",
            "description": "`bundle`, `description` or `issuer`."
          },
          "purge": {
            "type": "boolean",
            "description": "Bundles only: also delete the bytes from this registry's store.\nThe moderation entry stays, so the same bytes are refused if\nre-uploaded; the chain commitment and the name are untouched."
          },
          "reason": {
            "type": ["string", "null"],
            "description": "Optional operator note, stored with the entry (audit trail)."
          }
        }
      },
      "PauseRequest": {
        "type": "object",
        "description": "Set the mint pause.",
        "required": ["paused"],
        "properties": {
          "paused": {
            "type": "boolean",
            "description": "`true` pauses the relay and metadata uploads on this instance;\n`false` resumes them."
          },
          "reason": {
            "type": ["string", "null"],
            "description": "Optional operator note, shown on the admin page and kept with the\ndecision."
          }
        }
      },
      "PauseResponse": {
        "type": "object",
        "description": "The current pause decision.",
        "required": ["paused"],
        "properties": {
          "paused": {
            "type": "boolean"
          },
          "reason": {
            "type": ["string", "null"]
          },
          "since": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "Unix time (seconds) of the last change, when known.",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}