# server/crates/api/src/admin.rs
//! Operator admin surface: moderation over HTTP, token-gated.
//!
//! Disabled by default — with no CACHET_ADMIN_TOKEN in the environment
//! every route here answers 404, indistinguishable from not existing.
//! With a token, requests carry `Authorization: Bearer <token>`, compared
//! in constant time. This is the one deliberate exception to
//! "moderation only over SSH": it stays availability-only (hide, never
//! alter), auditable and reversible, exactly like the CLI it mirrors.

use axum::Json;
use axum::extract::State;
use serde::{Deserialize, Serialize};
use utoipa::ToSchema;

use crate::AppState;
use crate::error::ApiError;
use crate::routes::{metadata_error, require_metadata_store};

/// Constant-time token comparison. Both sides are hashed to a fixed
/// 32 bytes first, so neither the mismatch position nor the length of the
/// configured token shapes the timing.
fn constant_time_eq(a: &[u8], b: &[u8]) -> bool {
    use sha2::{Digest, Sha256};
    let a = Sha256::digest(a);
    let b = Sha256::digest(b);
    a.iter()
        .zip(b.iter())
        .fold(0u8, |acc, (x, y)| acc | (x ^ y))
        == 0
}

fn authorize(state: &AppState, headers: &axum::http::HeaderMap) -> Result<(), ApiError> {
    let Some(expected) = &state.admin_token else {
        // No token configured → the surface does not exist.
        return Err(ApiError::AdminUnauthorized);
    };
    // RFC 7235: the auth-scheme is case-insensitive.
    let provided = headers
        .get(axum::http::header::AUTHORIZATION)
        .and_then(|value| value.to_str().ok())
        .and_then(|value| {
            let (scheme, token) = value.split_once(' ')?;
            scheme
                .eq_ignore_ascii_case("bearer")
                .then_some(token.trim())
        })
        .unwrap_or_default();
    if constant_time_eq(provided.as_bytes(), expected.as_bytes()) {
        Ok(())
    } else {
        Err(ApiError::AdminUnauthorized)
    }
}

#[derive(Debug, Deserialize, ToSchema)]
pub struct ModerationRequest {
    /// `bundle`, `description` or `issuer`.
    pub kind: String,
    /// Hex-encoded key: bundle sha256, asset id, or issuance validating key.
    pub key: String,
    /// Optional operator note, stored with the entry (audit trail).
    pub reason: Option<String>,
    /// Bundles only: also delete the bytes from this registry's store.
    /// The moderation entry stays, so the same bytes are refused if
    /// re-uploaded; the chain commitment and the name are untouched.
    #[serde(default)]
    pub purge: bool,
}

/// Set the mint pause.
#[derive(Debug, Deserialize, ToSchema)]
pub struct PauseRequest {
    /// `true` pauses the relay and metadata uploads on this instance;
    /// `false` resumes them.
    pub paused: bool,
    /// Optional operator note, shown on the admin page and kept with the
    /// decision.
    pub reason: Option<String>,
}

/// The current pause decision.
#[derive(Debug, Serialize, ToSchema)]
pub struct PauseResponse {
    pub paused: bool,
    pub reason: Option<String>,
    /// Unix time (seconds) of the last change, when known.
    pub since: Option<u64>,
}

#[derive(Debug, Serialize, ToSchema)]
pub struct ModerationEntryResponse {
    pub kind: String,
    pub key: String,
    pub reason: Option<String>,
    pub hidden_at: String,
    /// Bundles only: whether this registry still holds the bytes
    /// (`false` after a purge). `null` for other kinds.
    pub bytes_present: Option<bool>,
}

fn parse_request(
    body: &ModerationRequest,
) -> Result<(cachet_index::ModerationKind, Vec<u8>), ApiError> {
    let kind = cachet_index::ModerationKind::parse(&body.kind).ok_or(ApiError::Validation(
        cachet_domain::DomainError::InvalidMetadata {
            reason: "kind must be bundle, description or issuer",
        },
    ))?;
    let key = hex::decode(body.key.trim()).map_err(|_| {
        ApiError::Validation(cachet_domain::DomainError::InvalidMetadata {
            reason: "key must be hex",
        })
    })?;
    // Length is part of the identity: a truncated asset id would be stored
    // happily and match nothing, so the operator would see "hidden" while
    // the content stayed visible. Silent moderation failure is the worst
    // kind, so a wrong length is refused loudly.
    let expected = match kind {
        cachet_index::ModerationKind::Issuer => 33, // issuance validating key
        _ => 32,                                    // bundle sha256 / asset id
    };
    if key.len() != expected {
        return Err(ApiError::Validation(
            cachet_domain::DomainError::InvalidMetadata {
                reason: "key has the wrong length for this kind (32 bytes for bundle and                          description, 33 for an issuer key)",
            },
        ));
    }
    Ok((kind, key))
}

/// Hide a bundle, description or issuer from this registry's distribution.
#[utoipa::path(
    post,
    path = "/api/v1/admin/moderation",
    tag = "ops",
    request_body = ModerationRequest,
    responses(
        (status = 204, description = "Hidden"),
        (status = 404, description = "Admin surface disabled or bad token"),
    )
)]
pub(crate) async fn hide(
    State(state): State<AppState>,
    headers: axum::http::HeaderMap,
    Json(body): Json<ModerationRequest>,
) -> Result<axum::http::StatusCode, ApiError> {
    authorize(&state, &headers)?;
    let store = require_metadata_store(&state)?;
    let (kind, key) = parse_request(&body)?;
    store
        .moderation_hide(kind, &key, body.reason.as_deref())
        .await
        .map_err(metadata_error)?;
    tracing::info!(kind = body.kind, key = body.key, "admin: hidden");
    if body.purge {
        if kind != cachet_index::ModerationKind::Bundle {
            return Err(ApiError::Validation(
                cachet_domain::DomainError::InvalidMetadata {
                    reason: "purge applies to bundles only",
                },
            ));
        }
        let sha256: [u8; 32] = key
            .as_slice()
            .try_into()
            .expect("parse_request checked the bundle key length");
        let existed = store.delete(sha256).await.map_err(metadata_error)?;
        tracing::info!(key = body.key, existed, "admin: purged bundle bytes");
    }
    Ok(axum::http::StatusCode::NO_CONTENT)
}

/// Lift a moderation entry.
#[utoipa::path(
    delete,
    path = "/api/v1/admin/moderation",
    tag = "ops",
    request_body = ModerationRequest,
    responses(
        (status = 204, description = "Unhidden (or was not hidden)"),
        (status = 404, description = "Admin surface disabled or bad token"),
    )
)]
pub(crate) async fn unhide(
    State(state): State<AppState>,
    headers: axum::http::HeaderMap,
    Json(body): Json<ModerationRequest>,
) -> Result<axum::http::StatusCode, ApiError> {
    authorize(&state, &headers)?;
    let store = require_metadata_store(&state)?;
    let (kind, key) = parse_request(&body)?;
    store
        .moderation_unhide(kind, &key)
        .await
        .map_err(metadata_error)?;
    tracing::info!(kind = body.kind, key = body.key, "admin: unhidden");
    Ok(axum::http::StatusCode::NO_CONTENT)
}

/// Every moderation entry, for the operator's audit view.
#[utoipa::path(
    get,
    path = "/api/v1/admin/moderation",
    tag = "ops",
    responses(
        (status = 200, body = Vec<ModerationEntryResponse>),
        (status = 404, description = "Admin surface disabled or bad token"),
    )
)]
pub(crate) async fn list(
    State(state): State<AppState>,
    headers: axum::http::HeaderMap,
) -> Result<Json<Vec<ModerationEntryResponse>>, ApiError> {
    authorize(&state, &headers)?;
    let store = require_metadata_store(&state)?;
    let entries = store.moderation_list().await.map_err(metadata_error)?;
    let mut responses = Vec::with_capacity(entries.len());
    for entry in entries {
        // A hidden bundle either still sits on disk or was purged; the
        // operator's view should say which.
        let bytes_present = if entry.kind == cachet_index::ModerationKind::Bundle.as_str() {
            match hex::decode(&entry.key)
                .ok()
                .and_then(|key| <[u8; 32]>::try_from(key.as_slice()).ok())
            {
                Some(sha256) => Some(store.get(sha256).await.map_err(metadata_error)?.is_some()),
                None => None,
            }
        } else {
            None
        };
        responses.push(ModerationEntryResponse {
            kind: entry.kind,
            key: entry.key,
            reason: entry.reason,
            hidden_at: entry.hidden_at,
            bytes_present,
        });
    }
    Ok(Json(responses))
}

/// The current pause decision.
#[utoipa::path(
    get,
    path = "/api/v1/admin/pause",
    tag = "ops",
    responses(
        (status = 200, body = PauseResponse),
        (status = 404, description = "Admin surface disabled or bad token"),
    )
)]
pub(crate) async fn pause_get(
    State(state): State<AppState>,
    headers: axum::http::HeaderMap,
) -> Result<Json<PauseResponse>, ApiError> {
    authorize(&state, &headers)?;
    let persisted = match &state.metadata {
        Some(store) => crate::pause_state_at_boot(store.as_ref()).await,
        None => crate::PauseState::default(),
    };
    Ok(Json(PauseResponse {
        // The live flag is the truth; the store adds the note and the time.
        paused: state
            .mints_paused
            .load(std::sync::atomic::Ordering::Relaxed),
        reason: persisted.reason,
        since: persisted.since,
    }))
}

/// Pause or resume minting through this instance. Effective on the next
/// request; persisted so a restart keeps it. The chain is not involved:
/// a paused instance simply stops relaying and storing.
#[utoipa::path(
    put,
    path = "/api/v1/admin/pause",
    tag = "ops",
    request_body = PauseRequest,
    responses(
        (status = 200, body = PauseResponse),
        (status = 404, description = "Admin surface disabled or bad token"),
    )
)]
pub(crate) async fn pause_set(
    State(state): State<AppState>,
    headers: axum::http::HeaderMap,
    Json(body): Json<PauseRequest>,
) -> Result<Json<PauseResponse>, ApiError> {
    authorize(&state, &headers)?;
    let reason = body
        .reason
        .as_deref()
        .map(str::trim)
        .filter(|text| !text.is_empty())
        .map(|text| text.chars().take(200).collect::<String>());
    let decision = crate::PauseState {
        paused: body.paused,
        reason: reason.clone(),
        since: std::time::SystemTime::now()
            .duration_since(std::time::UNIX_EPOCH)
            .ok()
            .map(|elapsed| elapsed.as_secs()),
    };
    // Flip first: the decision must hold even if persisting fails.
    state
        .mints_paused
        .store(body.paused, std::sync::atomic::Ordering::Relaxed);
    if let Some(store) = &state.metadata {
        let text = serde_json::to_string(&decision).expect("pause state serializes");
        if let Err(error) = store.setting_set(crate::MINTS_PAUSED_SETTING, &text).await {
            tracing::warn!(%error, "admin: pause decision applied but not persisted");
        }
    }
    tracing::info!(paused = body.paused, "admin: mint pause changed");
    Ok(Json(PauseResponse {
        paused: body.paused,
        reason,
        since: decision.since,
    }))
}

#[cfg(test)]
mod tests {
    use super::constant_time_eq;

    #[test]
    fn constant_time_eq_matches_semantics() {
        assert!(constant_time_eq(b"secret", b"secret"));
        assert!(!constant_time_eq(b"secret", b"secreT"));
        assert!(!constant_time_eq(b"secret", b"secret2"));
        assert!(!constant_time_eq(b"", b"x"));
        assert!(constant_time_eq(b"", b""));
    }
}